Add files via upload
Named Entity Recognition - NLP based

Pre Trained Model based on python. Named Entity Recognition (NER). NER is a part of natural language processing (NLP) and information retrieval (IR). The task in NER is to find the entity-type of words. Entities can, for example, be locations, time expressions or names. In given dataset, there are some job descriptions. My task is to categorize the keywords inside the dataset into 4 categories - Industry, Job title, Department, and Skills. On seeing the dataset they are JSON and there are direct key-value pairs to get some of the categories. I’ve need to specifically consider them as text documents and use NER to perform the categorization.